Add elasticsearch integration test
